requirements.txt
requests
openai
python-dotenv
csv